requirements.txt
pandas
requests
matplotlib
python-dotenv
SQLAlchemy
